+++ src/racket/gc2/newgc.h
@@ -13,8 +13,8 @@ typedef struct mpage {
   struct mpage *next;
   struct mpage *prev;
   void *addr;
-  unsigned long previous_size; /* for med page, place to search for available block; for jit nursery, allocated size */
-  unsigned long size; /* big page size, med page element size, or nursery starting point */
+  uintptr_t previous_size; /* for med page, place to search for available block; for jit nursery, allocated size */
+  uintptr_t size; /* big page size, med page element size, or nursery starting point */
 /*
   unsigned char generation    :1;
   unsigned char back_pointers :1;
@@ -45,21 +45,21 @@ typedef struct Gen0 {
   struct mpage *curr_alloc_page;
   struct mpage *pages;
   struct mpage *big_pages;
-  unsigned long current_size;
-  unsigned long max_size;
-  unsigned long page_alloc_size;
+  uintptr_t current_size;
+  uintptr_t max_size;
+  uintptr_t page_alloc_size;
 } Gen0;
 
 typedef struct MsgMemory {
   struct mpage *pages;
   struct mpage *big_pages;
-  unsigned long size;
+  uintptr_t size;
 } MsgMemory;
 
 typedef struct Allocator {
   Gen0 savedGen0;
-  unsigned long saved_alloc_page_ptr;
-  unsigned long saved_alloc_page_end;
+  uintptr_t saved_alloc_page_ptr;
+  uintptr_t saved_alloc_page_end;
 } Allocator;
 
 typedef struct MarkSegment {
@@ -78,22 +78,22 @@ typedef struct AccountHook {
   int type;
   void *c1;
   void *c2;
-  unsigned long amount;
+  uintptr_t amount;
   struct AccountHook *next;
 } AccountHook;
 
 typedef struct OTEntry {
   void *originator;
   void **members;
-  unsigned long memory_use;
-  unsigned long single_time_limit;
-  unsigned long super_required;
+  uintptr_t memory_use;
+  uintptr_t single_time_limit;
+  uintptr_t super_required;
   char limit_set;
   char required_set;
 } OTEntry;
 
 typedef struct Range {
-  unsigned long start, len;
+  uintptr_t start, len;
   struct Range *left, *right, *prev, *next;
 } Range;
 
@@ -101,22 +101,22 @@ typedef struct Page_Range {
   Range *range_root;
   Range *range_start;
   void *range_alloc_block;
-  unsigned long range_alloc_size;
-  unsigned long range_alloc_used;
+  uintptr_t range_alloc_size;
+  uintptr_t range_alloc_used;
 } Page_Range;
 
 typedef struct {
   char *start;
-  long len;
+  intptr_t len;
   short age;
   short zeroed;
 } AllocCacheBlock;
 
 #ifdef MZ_USE_PLACES
 typedef struct NewGCMasterInfo {
-  unsigned long size;
-  unsigned long alive;
-  unsigned long ready;
+  uintptr_t size;
+  uintptr_t alive;
+  uintptr_t ready;
   void **signal_fds;
   mzrt_rwlock *cangc;
   mzrt_sema *wait_sema;
@@ -149,19 +149,19 @@ typedef struct NewGC {
   Fnl *last_in_queue;
 
   struct NewGC *primoridal_gc;
-  unsigned long max_heap_size;
-  unsigned long max_pages_in_heap;
-  unsigned long max_pages_for_use;
-  unsigned long used_pages;
-  unsigned long actual_pages_size;
+  uintptr_t max_heap_size;
+  uintptr_t max_pages_in_heap;
+  uintptr_t max_pages_for_use;
+  uintptr_t used_pages;
+  uintptr_t actual_pages_size;
   void (*unsafe_allocation_abort)(struct NewGC *);
-  unsigned long memory_in_use; /* the amount of memory in use */
+  uintptr_t memory_in_use; /* the amount of memory in use */
 
   /* blame the child thread infos */
   GC_Thread_Info *thread_infos;
 
   mpage *release_pages;
-  unsigned long stack_base;
+  uintptr_t stack_base;
   int dumping_avoid_collection; /* dumping coutner flag */
 
   unsigned char generations_available        :1;
@@ -185,20 +185,20 @@ typedef struct NewGC {
   AccountHook *hooks;
 
 
-  unsigned long number_of_gc_runs;
+  uintptr_t number_of_gc_runs;
   unsigned int since_last_full;
-  unsigned long last_full_mem_use;
+  uintptr_t last_full_mem_use;
 
   /* These collect information about memory usage, for use in GC_dump. */
-  unsigned long peak_memory_use;
-  unsigned long num_minor_collects;
-  unsigned long num_major_collects;
+  uintptr_t peak_memory_use;
+  uintptr_t num_minor_collects;
+  uintptr_t num_major_collects;
   
   /* THREAD_LOCAL variables that need to be saved off */
   MarkSegment  *saved_mark_stack;
   void         *saved_GC_variable_stack;
-  unsigned long saved_GC_gen0_alloc_page_ptr;
-  unsigned long saved_GC_gen0_alloc_page_end;
+  uintptr_t saved_GC_gen0_alloc_page_ptr;
+  uintptr_t saved_GC_gen0_alloc_page_end;
   /* Distributed GC over places info */
 #ifdef MZ_USE_PLACES
   int           place_id;
@@ -211,8 +211,8 @@ typedef struct NewGC {
   /* Callbacks */
   void (*GC_collect_start_callback)(void);
   void (*GC_collect_end_callback)(void);
-  void (*GC_collect_inform_callback)(int major_gc, long pre_used, long post_used);
-  unsigned long (*GC_get_thread_stack_base)(void);
+  void (*GC_collect_inform_callback)(int major_gc, intptr_t pre_used, intptr_t post_used);
+  uintptr_t (*GC_get_thread_stack_base)(void);
 
   GC_Immobile_Box *immobile_boxes;
