+++ src/racket/gc2/stack_comp.c
@@ -1,7 +1,7 @@
 
 #include "mzconfig.h"
 #if STACK_DIRECTION > 0
-# define SHALLOWER_STACK_ADDRESS(a, b) ((unsigned long)a > (unsigned long)b)
+# define SHALLOWER_STACK_ADDRESS(a, b) ((uintptr_t)a > (uintptr_t)b)
 #else
-# define SHALLOWER_STACK_ADDRESS(a, b) ((unsigned long)a < (unsigned long)b)
+# define SHALLOWER_STACK_ADDRESS(a, b) ((uintptr_t)a < (uintptr_t)b)
 #endif
